requirements.txt
tensorflow-macos==2.16.2
numpy<2.0.0
opencv-python-headless
matplotlib
h5py